+++ test/Mergent.test.ts
@@ -1,5 +1,6 @@
 import Mergent from "../src/Mergent";
 import RequestValidator from "../src/RequestValidator";
+import Schedules from "../src/resources/Schedules";
 import Tasks from "../src/resources/Tasks";
 
 const mergent = new Mergent("");
@@ -7,6 +8,12 @@ const mergent = new Mergent("");
 describe("#tasks", () => {
   test("is an instance of the Tasks resource", () => {
     expect(mergent.tasks).toBeInstanceOf(Tasks);
+  });
+});
+
+describe("#schedules", () => {
+  test("is an instance of the Schedules resource", () => {
+    expect(mergent.schedules).toBeInstanceOf(Schedules);
   });
 });